gyp 파일 접근 및 수정 중

+++ build.py
@@ -15,13 +15,13 @@ def configLoad():
 
     """
 
-    global globalConfig
+    # global globalConfig
 
     content = open("sexy.json").read()
     print(json)
 
     stringIO = StringIO(content)
-    globalConfig = json.load(stringIO)
+    return json.load(stringIO)
 
 
 def main():
@@ -29,9 +29,7 @@ def main():
     build
 
     """
-    configLoad()
-
-    print(globalConfig["libPath"])
+    globalConfig = configLoad()
 
     suffix = globalConfig["suffix"]
     libPath = globalConfig["libPath"]
@@ -43,16 +41,51 @@ def main():
     #     #     if regexp.search(content):
     #     #         print(f)
     #
-        print(f)
+
+        # print(f)
         splitext = os.path.splitext(f)
         source = libPath + f
         target = libPath + splitext[0] + suffix + splitext[1]
 
-        shutil.move(source, target)
-        print(f +" move "+ target)
+        if os.path.exists(source):
+            shutil.move(source, target)
+            print(f +" move "+ target)
 
 
     # node.gyp 변경
+    gyp = globalConfig["gyp"]
+
+    old_gyp = ""
+    for line in open(gyp["path"]):
+        old_gyp += line
+
+    # lib list
+    start = old_gyp.find(gyp["prefix"])
+    end = old_gyp.find(gyp["suffix"], start)
+
+    libList = list(old_gyp)[start:end-4]
+    libText = "".join(libList).split("\n")
+    print(  libText )
+
+    new_gyp = ""
+    for line in libText:
+        # print (line)
+        if line.__contains__("'lib/"):
+            start = line.rfind("''lib/")
+            end = line.find("',", start)
+
+            func = list(line)[start + 7:end]
+
+            new_gyp += ("\t'lib/" + "".join(func) + "',\n")
+
+            # print("".join(func))
+        #     new_gyp += "".join(func)
+
+
+        else:
+            new_gyp += line +"\n"
+
+    print("".join(new_gyp))
 
     # source 내용 변경
